Move SourceTextDiffer and associated types to Workspaces
