+++ src/Razor/benchmarks/Microsoft.AspNetCore.Razor.Microbenchmarks/TextDifferencing/SourceTextDifferBenchmark.cs
@@ -4,11 +4,11 @@
 using System;
 using System.Linq;
 using BenchmarkDotNet.Attributes;
-using Microsoft.AspNetCore.Razor.LanguageServer;
 using Microsoft.AspNetCore.Razor.PooledObjects;
+using Microsoft.AspNetCore.Razor.TextDifferencing;
 using Microsoft.CodeAnalysis.Text;
 
-namespace Microsoft.AspNetCore.Razor.Microbenchmarks.Serialization;
+namespace Microsoft.AspNetCore.Razor.Microbenchmarks.TextDifferencing;
 
 public class SourceTextDifferBenchmark
 {
