+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/TextDifferencing/DiffEdit.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT license. See License.txt in the project root for license information.
 
-namespace Microsoft.AspNetCore.Razor.LanguageServer;
+namespace Microsoft.AspNetCore.Razor.TextDifferencing;
 
 internal readonly struct DiffEdit
 {
